requirements.txt
Flask>=2.3
flask-cors>=4.0.0
PyJWT>=2.8.0
bcrypt>=4.0.1
pydantic>=2.7.0
firebase-admin>=6.5.0
gensim>=4.3.2
konlpy>=0.6.0
JPype1>=1.4.1
scikit-learn>=1.4.0
python-dotenv>=1.0.1
pandas>=2.2.1
numpy>=1.26.4
